Use space when imploding scope array
I was getting an invalid scope field error when using your eloquent-oauth package. 
The problem was that when you implode the array with a '+' symbol and then use urlencode on that string it will convert the '+' to '%2B'. Instagram expects the scope parameter to contain a '+' symbol. A space will be converted to a '+' when using urlencode